+++ src/data/GraphProblems.kt
@@ -4,6 +4,7 @@ data class GraphCfg(
 	val graph: Graph,
 	val k: Int = 0,
 	val edgeWeight: List<Double> = emptyList(),
+	val edgeColor: List<String> = emptyList(),
 ){
 	override fun toString(): String = graph.toString()
 
@@ -16,6 +17,7 @@ data class GraphCfg(
 				graph = Graph.new(data["vertices"], data["edges"], isDirected),
 				k = data["k"].parseInt(),
 				edgeWeight = data["edgeWeight"].toDoubleList(),
+				edgeColor = data["edgeColor"].toStringList(),
 			)
 		}
 	}