+++ src/problem/Problems.kt
@@ -9,6 +9,7 @@ const val CliqueCover = "clique_cover"
 const val DominatingSet = "dominating_set"
 const val EdgeColoring = "edge_coloring"
 const val EdgeCover = "edge_cover"
+const val GraphMatching = "graph_matching"
 const val Interval = "interval"
 const val KCut = "k_cut"
 const val Knapsack = "knapsack"
@@ -32,6 +33,7 @@ val Creator: Map<String, (String, Int) -> Problem?>  = mapOf(
 	DominatingSet to ::newDominatingSet,
 	EdgeColoring to ::newEdgeColoring,
 	EdgeCover to ::newEdgeCover,
+	GraphMatching to ::newGraphMatching,
 	Interval to ::newInterval,
 	KCut to ::newKCut,
 	Knapsack to ::newKnapsack,